requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
pydantic==2.9.2
